backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
anthropic>=0.28.0
deepgram-sdk>=3.2.0
websockets>=12.0
httpx>=0.27.0
livekit>=0.11.0
livekit-api>=0.6.0
google-cloud-texttospeech>=2.16.0
supabase>=2.4.0
pydantic>=2.0.0
